% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example.R
\name{Clean.measurement.for.prism.help}
\alias{Clean.measurement.for.prism.help}
\title{Help for Clean.measurement.for.prism}
\usage{
Clean.measurement.for.prism.help()
}
\description{
Copies the example input / output
into the current working directory.
}
