% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_Functions.R
\name{Clean.measurement.for.prism}
\alias{Clean.measurement.for.prism}
\title{Transformation for df when doing PRISM input}
\usage{
Clean.measurement.for.prism(df, outname = "Cleaned.Prism.input")
}
\description{
Transformation for df when doing PRISM input
}
