% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics_functions.R
\name{cox_univar_fun}
\alias{cox_univar_fun}
\title{Univariate Cox Regression (Categorical Predictor)}
\usage{
cox_univar_fun(data_fun, res_time, res_event, var)
}
\arguments{
\item{data_fun}{Data frame containing survival data.}

\item{res_time}{Character. Name of survival time variable.}

\item{res_event}{Character. Name of event indicator (1 = event, 0 = censored).}

\item{var}{Character. Categorical predictor variable.}

\item{dig}{Rounding digits.}
}
\value{
A matrix of HRs, CIs, and p-values.
}
\description{
Fits a univariate Cox proportional hazards model for time-to-event data
using a categorical predictor. Reports hazard ratios (HR), 95\% CI,
Wald p-values for 2-category variables, and likelihood ratio test (LRT)
results for variables with >2 categories.
}
\examples{
\dontrun{
cox_univar_fun(mydata, "OS_time", "OS_event", "Stage")
}

}
