% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{logist_multi_fun_cts}
\alias{logist_multi_fun_cts}
\title{Multivariable Logistic Regression (Continuous Predictor)}
\usage{
logist_multi_fun_cts(
  var_model,
  var,
  response,
  data_fun,
  multi_var_all = multi_var_all
)
}
\arguments{
\item{var_model}{Character vector of variables included in the multivariable model.}

\item{var}{Character. The specific categorical variable to extract results for.}

\item{response}{Character. Name of binary outcome variable.}

\item{data_fun}{Data frame containing variables.}

\item{multi_var_all}{Character vector of full model candidate variables.
Used to determine whether \code{var} is included in \code{var_model}.}
}
\value{
A single-row matrix containing adjusted OR, CI, and p-value.
}
\description{
Extracts adjusted odds ratio for a continuous variable from a
multivariable logistic regression model.
}
\examples{
\dontrun{
logist_multi_fun_cts(
  var_model = c("Age", "Gender"),
  var       = "Age",
  response  = "Y",
  data_fun  = mydata
)
}

}
