% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Note_REF.R
\name{modb.update}
\alias{modb.update}
\title{Update this MODBrowser package}
\usage{
modb.update(I_have_dataset = TRUE, forceInstall = FALSE)
}
\arguments{
\item{I_have_dataset}{Logical. Whether to preserve and restore the current
datasets while updating the package.}

\item{forceInstall}{Logical. Passed to \code{remotes::install_github()} to force a
reinstall from GitHub.}
}
\description{
Update this MODBrowser package, Automatic detaching and attaching current MODB.

If you have your own dataset, use: I_have_dataset=TRUE
It will keep your datasets and Datasets infomation.xlsx
}
