% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MODBrowser-package}
\alias{MODBrowser}
\alias{MODBrowser-package}
\title{MODBrowser package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Modularized tools for dataset indexing, gene ID/symbol resolution, searching, printing, and metadata binding. Includes full roxygen2, pkgdown, and GitHub Actions support.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ScottCalcalas/XZDBfunction}
  \item Report bugs at \url{https://github.com/ScottCalcalas/XZDBfunction/issues}
}

}
\keyword{internal}
