% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bioinfo_Build eniv.R
\name{xiaopei.build.bioinfo.envi}
\alias{xiaopei.build.bioinfo.envi}
\title{Build environment for Bioinformation analysis}
\usage{
xiaopei.build.bioinfo.envi()
}
\description{
Build environment for Bioinformation analysis
}
