% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XZDB_functions.R
\name{xiaopei.clean.file}
\alias{xiaopei.clean.file}
\title{Clean all files and folders inside a directory}
\usage{
xiaopei.clean.file(idx_dir)
}
\arguments{
\item{idx_dir}{Character. Directory to clean.}
}
\value{
Invisibly TRUE.
}
\description{
Create the directory if it doesn't exist. And make sure it being cleaned
}
\examples{
xiaopei.clean.file("IndexedData")
xiaopei.clean.file("Output")

}
