% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Note_REF.R
\name{XZ.update}
\alias{XZ.update}
\title{Update this XZDBfunction package}
\usage{
XZ.update(I_have_dataset = FALSE)
}
\description{
Update this XZDBfunction package, Automatic detaching and attaching current XZDB.

If you have your own dataset, use: I_have_dataset=TRUE
}
