% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example.R
\name{xzdb.nowDataset}
\alias{xzdb.nowDataset}
\title{Get the current using datasets}
\usage{
xzdb.nowDataset(ToName = "datasets")
}
\arguments{
\item{ToName}{Character. Destination folder name under the current working
directory. Default: \code{"datasets"}.}
}
\description{
Get the current using datasets
}
