% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Note_REF.R
\name{xzdb.update}
\alias{xzdb.update}
\title{Update this XZDBfunction package}
\usage{
xzdb.update(I_have_dataset = TRUE, forceInstall = FALSE)
}
\description{
Update this XZDBfunction package, Automatic detaching and attaching current XZDB.

If you have your own dataset, use: I_have_dataset=TRUE
It will keep your datasets and Datasets infomation.xlsx
}
